requirements.txt
requests==2.31.0
Pillow==10.2.0
python-dotenv==1.0.1
Flask==3.0.2
Flask-WTF==1.2.1
Flask-Compress==1.14
Flask-Caching==2.1.0
aiohttp==3.9.3
beautifulsoup4==4.12.2 